Revert show-source and show-doc unify
